Add support for controlled tabs, move state to hoc
